+++ types/CustomStack.h
@@ -56,8 +56,8 @@ namespace types {
       std::stack<T> stack_;
   };
 
-  typedef std::unique_ptr<CustomStack<int>> NumbersStackPtr;
-  typedef CustomStack<int> NumbersStack;
+  typedef std::unique_ptr<CustomStack<float>> NumbersStackPtr;
+  typedef CustomStack<float> NumbersStack;
   typedef std::unique_ptr<std::vector<std::string>> OperatorsListPtr;
   typedef std::vector<std::string> OperatorsList;
 }
